HR Churn Prediction Project
Human Resource Analysis
Companies want to improve and maintain a healthy relationship and environment for their employees. This can be achieved by recognizing and understanding the important factors that are associated with employee turnover and taking care of them might add into the productivity and growth of the employees. These insights can help managers for grabing the opportunity to take corrective steps to build and preserve their successful business.

In this notebook, we will work on the project that is based on understanding what factors contributed most to employee turnover and to create a model that can predict if a certain employee will leave the company or not.


Column Description
satisfacion_level: Showing satisfaction of a particular employee
last_evaluation: Showing last evaluation of a particular employee
number_project: Showing number of projects handled a particular employee
average_montly_hours: Showing the monthly hours that were spent the particular emloyee
time_spend_company: Shows the number of years spent by the particular employee in the company.
Work_accident: Showing an employee has whether been part of the company or not.
left: Tells either and employee has left the company or not. Shows two values 0= not left, 1= left
promotion_last_5years: Shows that the whether the employee has got any promotion in the last 5 years or not.
dept: Shows the departments
salary: Shows the salary type of the employee